Bmv2 multi-architecture support
This commit refactors bmv2 backend to support multiple pipelien
architectures. JsonConverter is refactored into a pass manager
which is more conformed to the rest of the compiler design.

Added files:
backend.cpp       - the new backend pass manager
convert*.cpp      - compiler passes for control, parser, deparser,
                    externs and headers
inferArchitecture - a new frontend pass to parse the architecture p4
                    into a data object which is then used by json
                    generator
model.cpp         - a v2model that oriented for p4-16 IR

TODO:
tov1model         - a pass to convert v2model to v1model to execute
                    on simple-switch
