+++ backends/bmv2/Makefile.am
@@ -22,19 +22,40 @@ p4c_bm2_ss_LDADD = libfrontend.la libp4ctoolkit.a
 p4c_bm2_ss_UNIFIED = \
 	backends/bmv2/bmv2.cpp \
 	backends/bmv2/analyzer.cpp \
-	backends/bmv2/jsonconverter.cpp \
 	backends/bmv2/midend.cpp \
-	backends/bmv2/lower.cpp
+	backends/bmv2/lower.cpp \
+	backends/bmv2/helpers.cpp \
+	backends/bmv2/expressionConverter.cpp \
+	backends/bmv2/convertControl.cpp \
+	backends/bmv2/convertDeparser.cpp \
+	backends/bmv2/convertParser.cpp \
+	backends/bmv2/convertExterns.cpp \
+	backends/bmv2/convertHeaders.cpp \
+	backends/bmv2/convertErrorCode.cpp \
+	backends/bmv2/backend.cpp
 
 noinst_HEADERS += \
 	backends/bmv2/analyzer.h \
-	backends/bmv2/jsonconverter.h \
 	backends/bmv2/lower.h \
-	backends/bmv2/midend.h
+	backends/bmv2/midend.h \
+	backends/bmv2/helpers.h \
+	backends/bmv2/expressionConverter.h \
+	backends/bmv2/convertControl.h \
+	backends/bmv2/convertDeparser.h \
+	backends/bmv2/convertParser.h \
+	backends/bmv2/convertExterns.h \
+	backends/bmv2/convertHeaders.h \
+	backends/bmv2/convertErrorCode.h \
+	backends/bmv2/backend.h
 
 cpplint_FILES += $(p4c_bm2_ss_UNIFIED) $(p4c_bm2_ss_NONUNIFIED)
 
 ir_DEF_FILES += $(srcdir)/backends/bmv2/bmv2.def
+
+noinst_LIBRARIES += libbackendbmv2.a
+libbackendbmv2_a_LIBADD = libfrontend.la
+libbackendbmv2_a_SOURCES = \
+	$(p4c_bm2_ss_SOURCES)
 
 # Tests
 -include bmv2tests.mk
