+++ backends/bmv2/convertExterns.h
@@ -1,0 +1,57 @@
+/*
+Copyright 2013-present Barefoot Networks, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+#ifndef _BACKENDS_BMV2_CONVERTEXTERNS_H_
+#define _BACKENDS_BMV2_CONVERTEXTERNS_H_
+
+#include "ir/ir.h"
+#include "lib/json.h"
+#include "helpers.h"
+#include "frontends/p4/typeMap.h"
+#include "frontends/common/resolveReferences/referenceMap.h"
+
+namespace BMV2 {
+
+class DoExternBlockConversion : public Inspector {
+    P4::ReferenceMap*    refMap;
+    P4::TypeMap*         typeMap;
+    ExpressionConverter* conv;
+    Util::JsonArray*     externs;
+ protected:
+    void addExternAttributes(const IR::Declaration_Instance* di, const IR::ExternBlock* block,
+                             Util::JsonArray* attributes);
+
+ public:
+    bool preorder(const IR::PackageBlock* b) override;
+    bool preorder(const IR::Declaration_Instance* decl) override;
+
+    explicit DoExternBlockConversion(P4::ReferenceMap* refMap, P4::TypeMap* typeMap,
+                                     ExpressionConverter* conv, Util::JsonArray* externs) :
+        refMap(refMap), typeMap(typeMap), conv(conv), externs(externs) {}
+};
+
+class ConvertExterns final : public PassManager {
+ public:
+    ConvertExterns(P4::ReferenceMap* refMap, P4::TypeMap* typeMap,
+                   ExpressionConverter* conv, Util::JsonArray* externs) {
+       passes.push_back(new DoExternBlockConversion(refMap, typeMap, conv, externs));
+       setName("ConvertExterns");
+    }
+};
+
+}  // namespace BMV2
+
+#endif
