+++ p4include/psa.p4
@@ -523,7 +523,6 @@ extern ActionSelector {
 }
 
 
-
 /* --------------------------------------------------------------------- */
 /// \TODO: is generating a new packet and sending it to the stream or is
 /// it adding a header to the current packet and sending it to the
@@ -557,17 +556,19 @@ parser Parser<H, M>(packet_in buffer, out H parsed_hdr, inout M user_meta,
 
 control VerifyChecksum<H, M>(in H hdr, inout M user_meta);
 
+@control
 control Ingress<H, M>(inout H hdr, inout M user_meta,
                       PacketReplicationEngine pre,
                       in  psa_ingress_input_metadata_t  istd,
                       out psa_ingress_output_metadata_t ostd);
-
+@control
 control Egress<H, M>(inout H hdr, inout M user_meta,
                      PacketReplicationEngine pre,
                      in  psa_egress_input_metadata_t  istd);
 
 control ComputeChecksum<H, M>(inout H hdr, inout M user_meta);
 
+@deparser
 control Deparser<H>(packet_out buffer, in H hdr);
